refactor: Header 컴포넌트 분리 및 구조 최적화
- Logo, Navigation, UserProfile 컴포넌트를 별도 파일로 분리
- Avatar, Badge 재사용 가능한 UI 컴포넌트 생성
- 팀 컨벤션에 맞게 파일명 PascalCase 적용
- 상수 UPPERCASE 네이밍 적용
